Adds new methods to sun.misc.Unsafe: copyMemory, staticFieldBase, staticFieldOffset.
	Change on 2017/04/24 by kstanger <kstanger@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=154048643

+++ jre_emul/Classes/java/lang/reflect/Field.h
@@ -83,7 +83,7 @@
 - (id<JavaLangReflectType>)getGenericType;
 
 // Should only be used by sun.misc.Unsafe.
-- (int)unsafeOffset;
+- (jlong)unsafeOffset;
 
 @end
